Many small bugfixes and changes Move constants to a file Add half move clock to gamestate Fix bug with half move clock Insufficient material draws Fix bug with EP hashing Bench: 4777642

+++ src/search.cpp
@@ -3,6 +3,7 @@
 #include "movegen.h"
 #include "searcher.h"
 #include "config.h"
+#include "constants.h"
 #include "movepicker.h"
 #include "wdl.h"
 
@@ -397,7 +398,7 @@ MoveEvaluation iterativeDeepening(Board board, ThreadInfo& thisThread, SearchPar
                 break;
 
             // Soft TM
-            if (softTime > 0 && sp.time.elapsed() >= softTime)
+            if (softTime > 0 && static_cast<i64>(sp.time.elapsed()) >= softTime)
                 break;
         }
     }
